requirements.txt
adafruit_io==2.7.0
beautifulsoup4==4.10.0
Flask==2.2.2
Flask_Cors==3.0.10
numpy==1.22.4
opencv_contrib_python==4.5.5.64
opencv_python==4.5.5.64
pyserial==3.5
Requests==2.28.2
scikit_learn==1.0.2
scipy==1.10.1
tracker==0.1.1
